Expose loading script
