+++ js/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dqbd/tiktoken",
-  "version": "1.0.0-alpha.1",
+  "version": "1.0.0-alpha.2",
   "description": "Javascript bindings for tiktoken",
   "license": "MIT",
   "scripts": {
